pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "kinastro"
version = "2.0.0"
description = "Multi-system astrology chart application (堅占星)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.52.2",
    "pyswisseph>=2.10.3.2",
    "plotly>=6.0.1",
    "svgwrite>=1.4.3",
    "fpdf2>=2.7.6",
]

[project.urls]
Repository = "https://github.com/kentang2017/kinastro"
